ai-service/requirements.txt
flask
google-generativeai
pillow
ultralytics>=8.3.0
opencv-python-headless
numpy
python-dotenv
tensorflow